Truetype fonts, bitmap fonts and FBOs work, but still require further testing.

+++ addons/ofxGLES2Renderer/gles2-bc/OpenGLES/OpenGLESContext.h
@@ -38,12 +38,11 @@ typedef int             GLclampx;
 
 /* GL types for handling large vertex buffer objects */
 
+//typedef int             GLintptr;
+//typedef int             GLsizeiptr;
 
-typedef int             GLintptr;
-typedef int             GLsizeiptr;
-
-//typedef long             GLintptr;
-//typedef long             GLsizeiptr;
+typedef long             GLintptr;
+typedef long             GLsizeiptr;
 
 /* Extensions */
 #define GL_OES_byte_coordinates           1
